requirements.txt
flask>=2.0
netmiko>=4.6.0
Jinja2>=3.0
PyYAML>=6.0
